[sync code]fix task manager cannot clear render process.

+++ standard/appspawn_service.c
@@ -158,13 +158,17 @@ static void SignalHandler(const struct signalfd_siginfo *siginfo)
     APPSPAWN_LOGI("SignalHandler signum %d", siginfo->ssi_signo);
     switch (siginfo->ssi_signo) {
         case SIGCHLD: {  // delete pid from app map
-            // nwebspawn will invoke waitpid and remove appinfo at GetRenderProcessTerminationStatus.
             pid_t pid;
             int status;
             while ((pid = waitpid(-1, &status, WNOHANG)) > 0) {
                 APPSPAWN_LOGI("SignalHandler pid %d status %d", pid, status);
 #ifdef REPORT_EVENT
                 PrintProcessExitInfo(pid, siginfo->ssi_uid, status);
+#endif
+#ifdef NWEB_SPAWN
+                // nwebspawn will invoke waitpid and remove appinfo at GetProcessTerminationStatusInner when
+                // GetProcessTerminationStatusInner is called before the parent process receives the SIGCHLD signal.
+                RecordRenderProcessExitedStatus(pid, status);
 #endif
                 RemoveAppInfo(pid);
             }
@@ -245,10 +249,15 @@ static void StartColdApp(AppSpawnClientExt *appProperty)
 }
 
 #ifdef NWEB_SPAWN
-static int GetRenderProcessTerminationStatus(int32_t pid, int *status)
+static int GetProcessTerminationStatusInner(int32_t pid, int *status)
 {
     if (status == NULL) {
         return -1;
+    }
+
+    if (GetRenderProcessTerminationStatus(pid, status) == 0) {
+        // this shows that the parent process has recived SIGCHLD signal.
+        return 0;
     }
 
     if (kill(pid, SIGKILL) != 0) {
@@ -267,7 +276,7 @@ static int GetRenderProcessTerminationStatus(int32_t pid, int *status)
 static void GetProcessTerminationStatus(AppSpawnClientExt *appProperty)
 {
     int exitStatus = 0;
-    int ret = GetRenderProcessTerminationStatus(appProperty->property.pid, &exitStatus);
+    int ret = GetProcessTerminationStatusInner(appProperty->property.pid, &exitStatus);
     if (ret) {
         SendResponse(appProperty, (char *)&ret, sizeof(ret));
     } else {
